+++ src/editor/composables/index.js
@@ -7,5 +7,4 @@ export * from './useHistory';
 export * from './useWindows';
 export * from './useSaveTemplate';
 export * from './useElementActions';
-export * from './useAutosave';
 export * from './useLocalStorage';
